shreemoy/requirements.txt
cffi==1.17.1
cryptography==43.0.3
fuzzywuzzy==0.18.0
Levenshtein==0.26.1
numpy==2.1.3
pandas==2.2.3
py4j==0.10.9.7
pycparser==2.22
PyMySQL==1.1.1
pyodbc==5.2.0
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
RapidFuzz==3.10.1
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
